完成大部分背包系统中各个ui中的交互，具体的交互流程还是有点混乱

+++ Assets/Scripe/Bean/AccouterJsonBean.cs
@@ -49,7 +49,7 @@ public class AccouterJsonBean
             }
             string[] type = attribute_type.Split(',');
             for (int i = 0; i < type.Length; i++) {
-                mAttribute[i].type = int.Parse(type[i]);
+                mAttribute[i].type = long.Parse(type[i]);
             }
         }
         return mAttribute;
